reorganized things to properly have definitions either inlined or put in cpp files
