+++ src/core/ignore_parser.hpp
@@ -8,15 +8,12 @@ namespace core {
 
 struct IgnoreParser_routines {
 	protected:
-		static const routine_t on_start_, on_stop_;
-};
-
-const routine_t IgnoreParser_routines::on_start_ = [](const token_t& t) {
-	ParseFeeder::stop();
-};
-
-const routine_t IgnoreParser_routines::on_stop_ = [](const token_t& t) {
-	ParseFeeder::go();
+		static constexpr const routine_t on_start_ = [](const token_t& t) {
+			ParseFeeder::stop();
+		};
+		static constexpr const routine_t on_stop_ = [](const token_t& t) {
+			ParseFeeder::go();
+		};
 };
 
 class IgnoreParser : public ParseWorker, private IgnoreParser_routines {
