+++ src/core/parse_worker.hpp
@@ -12,6 +12,7 @@ namespace core {
 
 class ParseWorker;
 
+using symbol_t = Symbol;
 using token_t = Token<symbol_t>;
 using routine_t = void (*)(const token_t& t);
 using worker_t = ParseWorker;
@@ -61,7 +62,6 @@ class ParseWorker
 		class SymbolHandler
 		{
 			public:
-				using symbol_t = Symbol;
 				using symbols_t = std::unordered_set<symbol_t>;
 				using symbols_init_t = std::initializer_list<symbol_t>;
 
@@ -127,102 +127,6 @@ class ParseWorker
 		bool working_() const { return handler_i_ != 0 && !done_(); }
 		bool indefinite_() const { return iters_ == INF_ITERS; }
 };
-
-/*
- * Processes based on the current handler.
- */
-void ParseWorker::proc(const token_t& t)
-{
-	if (done_()) {
-		return;
-	}
-
-	if (handler_().match(t.name)) {
-		handler_().handle_token(t);
-
-		handler_().reset_workers();
-
-		++handler_i_;
-
-		if (done_()) {
-			if (indefinite_() || itered_++ < iters_) {
-				rewind_();
-			}
-			return;
-		}
-	}
-	else {
-		handler_().handle_workers(t);
-	}
-
-	return;
-}
-
-/*
- * Reset to the originally constructed state
- */
-void ParseWorker::reset()
-{
-	handler_().reset_workers();
-	rewind_();
-	itered_ = 0;
-}
-
-/*
- * Match to a symbol based on whether or not it is contained within symbols_;
- * if neg_ is true, this will match when the symbol is not contained.
- */
-bool ParseWorker::SymbolHandler::match(symbol_t s)
-{
-	return !neg_ ?
-		symbols_.find(s) != symbols_.end() : symbols_.find(s) == symbols_.end();
-}
-
-/*
- * Execute a routine by the on_symbol_ function pointer, if present.
- */
-void ParseWorker::SymbolHandler::handle_token(const token_t& t)
-{
-	if (on_symbol_) {
-		on_symbol_(t);
-	}
-}
-
-/*
- * Process all contained ParseWorker objects when none are "working";
- * if one is, process only that one until it is done.
- */
-void ParseWorker::SymbolHandler::handle_workers(const token_t& t)
-{
-	if (working_) {
-		working_->proc(t);
-		if (!working_->working_()) {
-			working_ = nullptr;
-			for (worker_t& p : workers_) {
-				p.reset();
-			}
-		}
-	}
-	if (!working_) {
-		for (worker_t& p : workers_) {
-			p.proc(t);
-			if (p.working_()) {
-				working_ = &p;
-				break;
-			}
-		}
-	}
-}
-
-/*
- * Reset all contained ParseWorker objects
- */
-void ParseWorker::SymbolHandler::reset_workers()
-{
-	for (worker_t& p : workers_) {
-		p.reset();
-	}
-}
 
 } // namespace core
 } // namespace docgen
