+++ src/core/tags_parser.hpp
@@ -9,13 +9,11 @@ namespace core {
 struct TagParser_routines
 {
 	protected:
-		static const routine_t on_tag_;
-};
-
-const routine_t TagParser_routines::on_tag_ = [](const token_t& t) {
-	ParseFeeder::at(t.content);
-	ParseFeeder::go();
-	ParseFeeder::skip();
+		static constexpr const routine_t on_tag_  = [](const token_t& t) {
+			ParseFeeder::at(t.content);
+			ParseFeeder::go();
+			ParseFeeder::skip();
+		};
 };
 
 class TagParser : public ParseWorker, private TagParser_routines
